// frontend/src/app/models/enums/notification-type.enum.ts
export enum NotificationType {
  FRIENDSHIP_REQUEST = 'FRIENDSHIP_REQUEST',
  FRIENDSHIP_ACCEPTED = 'FRIENDSHIP_ACCEPTED',
  MEMBERSHIP_REQUEST = 'MEMBERSHIP_REQUEST',
  MEMBERSHIP_ACCEPTED = 'MEMBERSHIP_ACCEPTED',
  SCHEDULED_MEETING = 'SCHEDULED_MEETING',
  PEER_MESSAGE = 'PEER_MESSAGE',
  ROOM_MESSAGE = 'ROOM_MESSAGE',
  TEST = 'TEST'
}